+++ CHANGELOG.md
@@ -2,3 +2,5 @@ Changes:
 - Added chat message limit setting
 - Added close tab button
 - Move settings out of chat tabs
+- Add button icons
+- Remove message send button